refactor(front): divide orgs pages into separate folders
