+++ frontend/src/App.tsx
@@ -14,9 +14,9 @@ import StoragePage from './pages/Storage'
 import AccountPage from './pages/Account'
 import OrganizationsPage from './pages/Organizations'
 import { ProtectedRoute } from './components/ProtectedRoute'
-import OrgFilesPage from './pages/Orgs/OrgFilesPage'
-import OrgMembersPage from './pages/Orgs/OrgMembersPage'
-import OrgSettingsPage from './pages/Orgs/OrgSettingsPage'
+import OrgFilesPage from './pages/OrganizationFiles'
+import OrgMembersPage from './pages/OrganizationMembers'
+import OrgSettingsPage from './pages/OrganizationSettings'
 import StatusPage from './pages/Status'
 import ScrollToTop from './components/ScrollToTop'
 import AboutPage from './pages/About/About'
